% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunction.R
\name{classFunction}
\alias{classFunction}
\title{Helper function to build the class vector}
\usage{
classFunction(datasetObject, column, diseaseTerms)
}
\arguments{
\item{datasetObject}{the Dataset object to build a class vector for}

\item{column}{column from the $pheno slot to look for the disease terms}

\item{diseaseTerms}{a list of terms identifying the disease samples}
}
\value{
returns a Dataset object that has a class vector inserted
}
\description{
Helper function to build the class vector
}
\details{
Based on a defined set of disease terms, builds a class vector.
}
\examples{
classObj <- classFunction(tinyMetaObject$originalData$Whole.Blood.Study.1, 
   column="group", diseaseTerms=c("Disease"))
}
\author{
Winston A. Haynes
}
