% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconutMetaIntegrator.R
\name{coconutMetaIntegrator}
\alias{coconutMetaIntegrator}
\title{A wrapper function to run COCONUT on the MetaIntegrator objects.}
\usage{
coconutMetaIntegrator(metaObject, ...)
}
\arguments{
\item{metaObject}{a MetaIntegrator formatted Meta object.}

\item{...}{pass along arguments to COCONUT}
}
\value{
Results from COCONUT analysis on the MetaIntegrator object
}
\description{
A wrapper function to run COCONUT on the MetaIntegrator objects.
}
\author{
Winston A. Haynes
}
