% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSymbolCorrection.R
\name{geneSymbolCorrection}
\alias{geneSymbolCorrection}
\title{Correct/update gene symbols in a metaObject}
\usage{
geneSymbolCorrection(metaObject)
}
\arguments{
\item{metaObject}{your metaObject}
}
\value{
A modified version of the input metaObject with updated gene symbols for each dataset in \code{metaObject$originalData}
}
\description{
The gene symbols in gene expression data are sometimes outdated or incorrect, so this function goes through
your metaObject and updates the symbols based on the HGNChelper package, as well as correcting some other known issues.
}
\examples{
tinyMetaObject = geneSymbolCorrection(tinyMetaObject)
}
\author{
Aditya Rao
}
