% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMostRecentFilter.R
\name{getMostRecentFilter}
\alias{getMostRecentFilter}
\title{Get name of most recent filter}
\usage{
getMostRecentFilter(metaObject)
}
\arguments{
\item{metaObject}{A meta object}
}
\value{
Name of the most recent filter
}
\description{
Given a \code{metaObject} this function will look through \code{$filterResults} 
for the most recent filter used and return the filter name.
}
\examples{
	getMostRecentFilter(tinyMetaObject)
}
\author{
Francesco Vallania
}
\keyword{attribute}
\keyword{utilities}
