% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunoStates.R
\name{immunoStatesDecov}
\alias{immunoStatesDecov}
\title{immunoStates deconvolution analysis on MetaIntegrator object(s)}
\usage{
immunoStatesDecov(metaObject)
}
\arguments{
\item{metaObject}{a MetaIntegrator formatted Meta object.}
}
\value{
Results from immunoStates on the MetaIntegrator object are 
stored in $immunoStates of the original Meta object
}
\description{
immunoStates deconvolution analysis on MetaIntegrator object(s)
}
\examples{
\dontrun{
# Example won't work on tinyMetaObject because it requires real gene names
# Download the needed datasets for processing. 
sleData <- getGEOData(c("GSE11909","GSE50635", "GSE39088"))

# Run immunoStates
immunoStatesEstimates <- immunoStateDecov(sleData)
}
}
\author{
Francesco Vallania
}
