% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Generates a Manhattan plot with effect size FDR as y-axis}
\usage{
manhattanPlot(metaObject)
}
\arguments{
\item{metaObject}{a Meta object which must have meta-analysis run}
}
\value{
Generates a Manhattan plot with effect size FDR as y-axis
}
\description{
Generates a Manhattan plot with effect size FDR as y-axis
}
\author{
Winston A. Haynes
}
