% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleROCPlot.R
\name{multipleROCPlot}
\alias{multipleROCPlot}
\title{Generate a plot with multiple ROC curves}
\usage{
multipleROCPlot(metaObject, filterObject, title = "title", size = 16)
}
\arguments{
\item{metaObject}{a Meta object which must have the $originalData populated}

\item{filterObject}{a MetaFilter object containing the signature genes that will be used for calculating the score}

\item{title}{title of the plot}

\item{size}{use this to easily increase or decrease the size of all the text in the plot}
}
\value{
Generates an ROC plot for all datasets
}
\description{
Generate a plot with multiple ROC curves
}
\examples{
multipleROCPlot(tinyMetaObject, filterObject = 
   tinyMetaObject$filterResults$pValueFDR0.05_es0_nStudies1_looaTRUE_hetero0)
}
