% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetOriginalData.R
\name{subsetOriginalData}
\alias{subsetOriginalData}
\title{Subset samples for a particular dataset}
\usage{
subsetOriginalData(datasetObject, keepMe)
}
\arguments{
\item{datasetObject}{the Dataset object to subset}

\item{keepMe}{either a binary vector for whether each sample should be 
in the subset or a list of names of samples to be in the subset}
}
\value{
returns a Dataset object that has been subsetted to the desired samples
}
\description{
Subset samples for a particular dataset
}
\details{
Subsets all relevant slots within the Dataset object to include only the 
desired samples.
}
\examples{
subsetObject <- subsetOriginalData(tinyMetaObject$originalData$Whole.Blood.Study.1, 
   keepMe= c("Sample 1", "Sample 13", "Sample 43"))
}
\author{
Winston A. Haynes
}
