% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryROCPlot.R
\name{summaryROCCalc}
\alias{summaryROCCalc}
\title{Calculate the summaryROC statistics}
\usage{
summaryROCCalc(metaObject, filterObject, bootstrapReps = 500)
}
\arguments{
\item{metaObject}{a Meta object which must have the $originalData populated}

\item{filterObject}{a MetaFilter object containing the signature genes that will be used for calculating the score}

\item{bootstrapReps}{number of bootstrap simulations to run for confidence interval on summary ROC}
}
\value{
Summary AUC statistics
}
\description{
Calculate the summaryROC statistics
}
\examples{
\dontrun{
summaryROCCalc(tinyMetaObject, filterObject = 
   tinyMetaObject$filterResults$pValueFDR0.05_es0_nStudies1_looaTRUE_hetero0)
}
}
